+++ setup.cfg
@@ -39,6 +39,7 @@ install_requires =
     pandas<2
     pathlib
     prefect>2
+    prefect-aws
     requests
     s3fs
     scipy
